Split mask classes into separate module
