+++ spectral_cube/tests/test_masks.py
@@ -3,7 +3,7 @@ import numpy as np
 from numpy.testing import assert_allclose
 from astropy.wcs import WCS
 
-from ..spectral_cube import SpectralCubeMask, SpectralCube, LazyMask, FunctionMask, CompositeMask
+from .. import SpectralCubeMask, SpectralCube, LazyMask, FunctionMask, CompositeMask
 
 
 def test_spectral_cube_mask():
@@ -13,51 +13,51 @@ def test_spectral_cube_mask():
 
     m = SpectralCubeMask(mask, mask_wcs)
 
-    data = np.arange(5).reshape((1,1,5))
+    data = np.arange(5).reshape((1, 1, 5))
     wcs = WCS()
 
-    assert_allclose(m.include(data, wcs), [[[0,1,1,0,1]]])
-    assert_allclose(m.exclude(data, wcs), [[[1,0,0,1,0]]])
-    assert_allclose(m._filled(data, wcs), [[[np.nan,1,2,np.nan,4]]])
-    assert_allclose(m._flattened(data, wcs), [1,2,4])
+    assert_allclose(m.include(data, wcs), [[[0, 1, 1, 0, 1]]])
+    assert_allclose(m.exclude(data, wcs), [[[1, 0, 0, 1, 0]]])
+    assert_allclose(m._filled(data, wcs), [[[np.nan, 1, 2, np.nan, 4]]])
+    assert_allclose(m._flattened(data, wcs), [1, 2, 4])
 
-    assert_allclose(m.include(data, wcs, view=(0,0,slice(1,4))), [1,1,0])
-    assert_allclose(m.exclude(data, wcs, view=(0,0,slice(1,4))), [0,0,1])
-    assert_allclose(m._filled(data, wcs, view=(0,0,slice(1,4))), [1,2,np.nan])
-    assert_allclose(m._flattened(data, wcs, view=(0,0,slice(1,4))), [1,2])
+    assert_allclose(m.include(data, wcs, view=(0, 0, slice(1, 4))), [1, 1, 0])
+    assert_allclose(m.exclude(data, wcs, view=(0, 0, slice(1, 4))), [0, 0, 1])
+    assert_allclose(m._filled(data, wcs, view=(0, 0, slice(1, 4))), [1, 2, np.nan])
+    assert_allclose(m._flattened(data, wcs, view=(0, 0, slice(1, 4))), [1, 2])
 
 
 def test_lazy_mask():
 
-    data = np.arange(5).reshape((1,1,5))
+    data = np.arange(5).reshape((1, 1, 5))
     wcs = WCS()
 
     m = LazyMask(lambda x: x > 2, data=data, wcs=wcs)
 
-    assert_allclose(m.include(data, wcs), [[[0,0,0,1,1]]])
-    assert_allclose(m.exclude(data, wcs), [[[1,1,1,0,0]]])
-    assert_allclose(m._filled(data, wcs), [[[np.nan,np.nan,np.nan,3,4]]])
-    assert_allclose(m._flattened(data, wcs), [3,4])
+    assert_allclose(m.include(data, wcs), [[[0, 0, 0, 1, 1]]])
+    assert_allclose(m.exclude(data, wcs), [[[1, 1, 1, 0, 0]]])
+    assert_allclose(m._filled(data, wcs), [[[np.nan, np.nan, np.nan, 3, 4]]])
+    assert_allclose(m._flattened(data, wcs), [3, 4])
 
-    assert_allclose(m.include(data, wcs, view=(0,0,slice(1,4))), [0,0,1])
-    assert_allclose(m.exclude(data, wcs, view=(0,0,slice(1,4))), [1,1,0])
-    assert_allclose(m._filled(data, wcs, view=(0,0,slice(1,4))), [np.nan,np.nan,3])
-    assert_allclose(m._flattened(data, wcs, view=(0,0,slice(1,4))), [3])
+    assert_allclose(m.include(data, wcs, view=(0, 0, slice(1, 4))), [0, 0, 1])
+    assert_allclose(m.exclude(data, wcs, view=(0, 0, slice(1, 4))), [1, 1, 0])
+    assert_allclose(m._filled(data, wcs, view=(0, 0, slice(1, 4))), [np.nan, np.nan, 3])
+    assert_allclose(m._flattened(data, wcs, view=(0, 0, slice(1, 4))), [3])
 
     # Now if we call with different data, the results for include and exclude
     # should *not* change.
 
-    data = (3-np.arange(5)).reshape((1,1,5))
+    data = (3 - np.arange(5)).reshape((1, 1, 5))
 
-    assert_allclose(m.include(data, wcs), [[[0,0,0,1,1]]])
-    assert_allclose(m.exclude(data, wcs), [[[1,1,1,0,0]]])
-    assert_allclose(m._filled(data, wcs), [[[np.nan,np.nan,np.nan,0,-1]]])
-    assert_allclose(m._flattened(data, wcs), [0,-1])
+    assert_allclose(m.include(data, wcs), [[[0, 0, 0, 1, 1]]])
+    assert_allclose(m.exclude(data, wcs), [[[1, 1, 1, 0, 0]]])
+    assert_allclose(m._filled(data, wcs), [[[np.nan, np.nan, np.nan, 0, -1]]])
+    assert_allclose(m._flattened(data, wcs), [0, -1])
 
-    assert_allclose(m.include(data, wcs, view=(0,0,slice(1,4))), [0,0,1])
-    assert_allclose(m.exclude(data, wcs, view=(0,0,slice(1,4))), [1,1,0])
-    assert_allclose(m._filled(data, wcs, view=(0,0,slice(1,4))), [np.nan,np.nan,0])
-    assert_allclose(m._flattened(data, wcs, view=(0,0,slice(1,4))), [0])
+    assert_allclose(m.include(data, wcs, view=(0, 0, slice(1, 4))), [0, 0, 1])
+    assert_allclose(m.exclude(data, wcs, view=(0, 0, slice(1, 4))), [1, 1, 0])
+    assert_allclose(m._filled(data, wcs, view=(0, 0, slice(1, 4))), [np.nan, np.nan, 0])
+    assert_allclose(m._flattened(data, wcs, view=(0, 0, slice(1, 4))), [0])
 
 
 def test_function_mask_incorrect_shape():
@@ -69,11 +69,11 @@ def test_function_mask_incorrect_shape():
 
     m = FunctionMask(threshold)
 
-    data = np.arange(5).reshape((1,1,5))
+    data = np.arange(5).reshape((1, 1, 5))
     wcs = WCS()
 
     with pytest.raises(ValueError) as exc:
-        m.include(data, wcs, view=(0,0,slice(1,4)))
+        m.include(data, wcs, view=(0, 0, slice(1, 4)))
     assert exc.value.args[0] == "Function did not return mask with correct shape - expected (3,), got (1, 1, 5)"
 
 
@@ -84,33 +84,33 @@ def test_function_mask():
 
     m = FunctionMask(threshold)
 
-    data = np.arange(5).reshape((1,1,5))
+    data = np.arange(5).reshape((1, 1, 5))
     wcs = WCS()
 
-    assert_allclose(m.include(data, wcs), [[[0,0,0,1,1]]])
-    assert_allclose(m.exclude(data, wcs), [[[1,1,1,0,0]]])
-    assert_allclose(m._filled(data, wcs), [[[np.nan,np.nan,np.nan,3,4]]])
-    assert_allclose(m._flattened(data, wcs), [3,4])
+    assert_allclose(m.include(data, wcs), [[[0, 0, 0, 1, 1]]])
+    assert_allclose(m.exclude(data, wcs), [[[1, 1, 1, 0, 0]]])
+    assert_allclose(m._filled(data, wcs), [[[np.nan, np.nan, np.nan, 3, 4]]])
+    assert_allclose(m._flattened(data, wcs), [3, 4])
 
-    assert_allclose(m.include(data, wcs, view=(0,0,slice(1,4))), [0,0,1])
-    assert_allclose(m.exclude(data, wcs, view=(0,0,slice(1,4))), [1,1,0])
-    assert_allclose(m._filled(data, wcs, view=(0,0,slice(1,4))), [np.nan,np.nan,3])
-    assert_allclose(m._flattened(data, wcs, view=(0,0,slice(1,4))), [3])
+    assert_allclose(m.include(data, wcs, view=(0, 0, slice(1, 4))), [0, 0, 1])
+    assert_allclose(m.exclude(data, wcs, view=(0, 0, slice(1, 4))), [1, 1, 0])
+    assert_allclose(m._filled(data, wcs, view=(0, 0, slice(1, 4))), [np.nan, np.nan, 3])
+    assert_allclose(m._flattened(data, wcs, view=(0, 0, slice(1, 4))), [3])
 
     # Now if we call with different data, the results for include and exclude
     # *should* change.
 
-    data = (3-np.arange(5)).reshape((1,1,5))
+    data = (3 - np.arange(5)).reshape((1, 1, 5))
 
-    assert_allclose(m.include(data, wcs), [[[1,0,0,0,0]]])
-    assert_allclose(m.exclude(data, wcs), [[[0,1,1,1,1]]])
-    assert_allclose(m._filled(data, wcs), [[[3, np.nan,np.nan,np.nan,np.nan]]])
+    assert_allclose(m.include(data, wcs), [[[1, 0, 0, 0, 0]]])
+    assert_allclose(m.exclude(data, wcs), [[[0, 1, 1, 1, 1]]])
+    assert_allclose(m._filled(data, wcs), [[[3, np.nan, np.nan, np.nan, np.nan]]])
     assert_allclose(m._flattened(data, wcs), [3])
 
-    assert_allclose(m.include(data, wcs, view=(0,0,slice(0,3))), [1,0,0])
-    assert_allclose(m.exclude(data, wcs, view=(0,0,slice(0,3))), [0,1,1])
-    assert_allclose(m._filled(data, wcs, view=(0,0,slice(0,3))), [3,np.nan,np.nan])
-    assert_allclose(m._flattened(data, wcs, view=(0,0,slice(0,3))), [3])
+    assert_allclose(m.include(data, wcs, view=(0, 0, slice(0, 3))), [1, 0, 0])
+    assert_allclose(m.exclude(data, wcs, view=(0, 0, slice(0, 3))), [0, 1, 1])
+    assert_allclose(m._filled(data, wcs, view=(0, 0, slice(0, 3))), [3, np.nan, np.nan])
+    assert_allclose(m._flattened(data, wcs, view=(0, 0, slice(0, 3))), [3])
 
 
 def test_composite_mask():
@@ -126,23 +126,23 @@ def test_composite_mask():
 
     m = m1 & m2
 
-    data = np.arange(5).reshape((1,1,5))
+    data = np.arange(5).reshape((1, 1, 5))
     wcs = WCS()
 
-    assert_allclose(m.include(data, wcs), [[[0,1,1,0,0]]])
-    assert_allclose(m.exclude(data, wcs), [[[1,0,0,1,1]]])
-    assert_allclose(m._filled(data, wcs), [[[np.nan,1,2,np.nan,np.nan]]])
-    assert_allclose(m._flattened(data, wcs), [1,2])
+    assert_allclose(m.include(data, wcs), [[[0, 1, 1, 0, 0]]])
+    assert_allclose(m.exclude(data, wcs), [[[1, 0, 0, 1, 1]]])
+    assert_allclose(m._filled(data, wcs), [[[np.nan, 1, 2, np.nan, np.nan]]])
+    assert_allclose(m._flattened(data, wcs), [1, 2])
 
-    assert_allclose(m.include(data, wcs, view=(0,0,slice(1,4))), [1,1,0])
-    assert_allclose(m.exclude(data, wcs, view=(0,0,slice(1,4))), [0,0,1])
-    assert_allclose(m._filled(data, wcs, view=(0,0,slice(1,4))), [1, 2, np.nan])
-    assert_allclose(m._flattened(data, wcs, view=(0,0,slice(1,4))), [1, 2])
+    assert_allclose(m.include(data, wcs, view=(0, 0, slice(1, 4))), [1, 1, 0])
+    assert_allclose(m.exclude(data, wcs, view=(0, 0, slice(1, 4))), [0, 0, 1])
+    assert_allclose(m._filled(data, wcs, view=(0, 0, slice(1, 4))), [1, 2, np.nan])
+    assert_allclose(m._flattened(data, wcs, view=(0, 0, slice(1, 4))), [1, 2])
 
 
 def test_mask_logic():
 
-    data = np.arange(5).reshape((1,1,5))
+    data = np.arange(5).reshape((1, 1, 5))
     wcs = WCS()
 
     def threshold_1(data, wcs, view=()):
@@ -159,16 +159,16 @@ def test_mask_logic():
     m3 = FunctionMask(threshold_3)
 
     m = m1 & m2
-    assert_allclose(m.include(data, wcs), [[[0,1,1,1,0]]])
+    assert_allclose(m.include(data, wcs), [[[0, 1, 1, 1, 0]]])
 
     m = m1 | m2
-    assert_allclose(m.include(data, wcs), [[[1,1,1,1,1]]])
+    assert_allclose(m.include(data, wcs), [[[1, 1, 1, 1, 1]]])
 
     m = m1 | ~m2
-    assert_allclose(m.include(data, wcs), [[[0,1,1,1,1]]])
+    assert_allclose(m.include(data, wcs), [[[0, 1, 1, 1, 1]]])
 
     m = m1 & m2 & m3
-    assert_allclose(m.include(data, wcs), [[[0,1,0,1,0]]])
+    assert_allclose(m.include(data, wcs), [[[0, 1, 0, 1, 0]]])
 
     m = (m1 | m3) & m2
-    assert_allclose(m.include(data, wcs), [[[1,1,1,1,0]]])
+    assert_allclose(m.include(data, wcs), [[[1, 1, 1, 1, 0]]])
